requirements.txt
selenium==4.15.0
webdriver-manager==4.0.1
pytest==7.4.3
pytest-html==4.1.1
python-dotenv==1.0.0
pandas==2.1.3